+++ components/frontend/src/trend_table/MeasurementsRow.test.js
@@ -14,9 +14,9 @@ describe("MeasurementRow", () => {
             start: "2020-01-06T00:00:00+00:00",
             end: "2020-01-09T00:00:00+00:00",
             count: {
-                value: "value 1",
+                value: "1",
                 status: "near_target_met",
-                target: "target 1",
+                target: "1",
             }
         },
         {
@@ -24,9 +24,9 @@ describe("MeasurementRow", () => {
             start: "2020-01-04T00:00:00+00:00",
             end: "2020-01-06T00:00:00+00:00",
             count: {
-                value: "value 0",
+                value: "0",
                 status: "near_target_met",
-                target: "target 0",
+                target: "0",
             }
         },
     ]
@@ -44,8 +44,8 @@ describe("MeasurementRow", () => {
     
         expect(queryAllByText("testName").length).toBe(1) // measurement name cell
         expect(queryAllByText("?").length).toBe(1) // first date before first measurement
-        expect(queryAllByText("value 0").length).toBe(1) // two cells with measurements
-        expect(queryAllByText("value 1").length).toBe(1) // two cells with measurements
+        expect(queryAllByText("0").length).toBe(1) // two cells with measurements
+        expect(queryAllByText("1").length).toBe(1) // two cells with measurements
         expect(queryAllByText("testUnit").length).toBe(1)
       });
 
@@ -58,10 +58,23 @@ describe("MeasurementRow", () => {
         expect(queryAllByText("Measurement").length).toBe(1) // measurement name cell
         expect(queryAllByText("Target").length).toBe(1) // target name cell
         expect(queryAllByText("?").length).toBe(2) // first date before first measurement
-        expect(queryAllByText("value 0").length).toBe(1) // two cells with measurements
-        expect(queryAllByText("value 1").length).toBe(1) // two cells with measurements
-        expect(queryAllByText("target 0").length).toBe(1) // two cells with targets
-        expect(queryAllByText("target 1").length).toBe(1) // two cells with targets
-        expect(queryAllByText("testUnit").length).toBe(2)
+        expect(queryAllByText("0").length).toBe(2) // two cells with measurements
+        expect(queryAllByText("1").length).toBe(2) // two cells with measurements
+        expect(queryAllByText("testUnit").length).toBe(2) // measurement and target tow both have the unit
     });
+
+     it('Renders one single row with metric name, measurement values and minutes unit', () => {
+    
+        metric["unit"] = ""
+
+        const { queryAllByText } = render(
+          <table><tbody><MeasurementsRow metricType={{unit: "minutes"}} metricName="testName" metric={metric} measurements={measurements} dates={dates} /></tbody></table>
+        );
+    
+        expect(queryAllByText("testName").length).toBe(1) // measurement name cell
+        expect(queryAllByText("?").length).toBe(1) // first date before first measurement
+        expect(queryAllByText("0:00").length).toBe(1) // two cells with measurements
+        expect(queryAllByText("0:01").length).toBe(1) // two cells with measurements
+        expect(queryAllByText("hours").length).toBe(1)
+      });
 })